Implementa o VendedorService

+++ src/main/java/com/rosario/garagem_automotiva/dto/CadastroCarroDTO.java
@@ -4,9 +4,10 @@ import com.rosario.garagem_automotiva.entity.MarcaCarro;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
-public record CadastroCarroDTO(@NotNull Double valor,
+public record CadastroCarroDTO(@NotNull BigDecimal valor,
                                @NotBlank String modelo,
                                @NotNull MarcaCarro marcaCarro,
                                @NotNull int ano,
