+++ src/main/java/com/rosario/garagem_automotiva/dto/CarroDTO.java
@@ -4,10 +4,11 @@ import com.rosario.garagem_automotiva.entity.MarcaCarro;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record CarroDTO(@NotNull UUID uuid,
-                       @NotNull Double valor,
+                       @NotNull BigDecimal valor,
                        @NotBlank String modelo,
                        @NotNull MarcaCarro marcaCarro,
                        @NotNull int ano,
